move JobMetadataChange to samza-api
