+++ samza-api/src/main/java/org/apache/samza/job/JobMetadataChange.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.samza.job.metadata;
+package org.apache.samza.job;
 
 /**
  * Provides granularity into changes in job metadata (e.g. new job model, new config).
