+++ samza-core/src/main/java/org/apache/samza/job/metadata/JobCoordinatorMetadataManager.java
@@ -37,6 +37,7 @@ import org.apache.samza.config.Config;
 import org.apache.samza.coordinator.stream.CoordinatorStreamValueSerde;
 import org.apache.samza.coordinator.stream.messages.SetJobCoordinatorMetadataMessage;
 import org.apache.samza.job.JobCoordinatorMetadata;
+import org.apache.samza.job.JobMetadataChange;
 import org.apache.samza.job.model.JobModel;
 import org.apache.samza.metadatastore.MetadataStore;
 import org.apache.samza.metrics.Gauge;
